\name{EdgesBackgrandData}
\alias{EdgesBackgrandData}

\docType{data}

\title{The data for the background set of edges}
\description{
  The data for the background set of edges in the environment variable of the system.
}
\details{
  The data for the background set of edges are extracted from seven pathway 
  database (KEGG; Biocarta; Reactome; NCI; SPIKE; HumanCyc; Panther). 
}


\format{An environment variable}
\author{Junwei Han <hanjunwei1981@163.com>, Xinrui Shi<xinrui103@163.com> and Chunquan Li <lcqbio@163.com>}
\keyword{file}