\name{envData}
\alias{envData}
\docType{data}

\title{The variables in the environment variable \code{envData} of the system}
\description{
  The variables in the environment variable \code{envData} of the system.
}

\format{An environment variable}
\author{Junwei Han <hanjunwei1981@163.com>, Xinrui Shi<xinrui103@163.com> and Chunquan Li <lcqbio@163.com>}
\keyword{file}