\name{ESEA-internal}
\alias{initializeESEA}
\alias{Getenvir}
\title{ESEA internal functions}
\description{
  These functions internally called by other functions in the package \code{ESEA} and not meant to be called by the user.
}

\usage{

    initializeESEA()
    Getenvir(envData)
}

\author{Junwei Han <hanjunwei1981@163.com>, Xinrui Shi<xinrui103@163.com> and Chunquan Li <lcqbio@163.com>}

\keyword{internal}