\name{ExampleData}
\alias{dataset}
\alias{class.labels}
\alias{controlcharactor}
\alias{Results}

\docType{data}

\title{The example data in the environment variable of the system}
\description{
  The example data in the environment variable of the system.
}
\details{
  The example data includes the variable \code{dataset}, \code{class.labels},\code{controlcharactor},\code{Results}. 
}

\format{An environment variable}
\author{Junwei Han <hanjunwei1981@163.com>, Xinrui Shi<xinrui103@163.com> and Chunquan Li <lcqbio@163.com>}
\keyword{file}