\name{PathwayEdgeData}
\alias{PathwayEdgeData}

\docType{data}

\title{The edge sets of pathways}

\description{
  The edge sets of pathways for the seven pathway database (KEGG; Biocarta; Reactome; NCI; SPIKE; HumanCyc; Panther) in the environment variable of the system.
}
\details{
  For each pathway in the seven pathway database (KEGG; Biocarta; Reactome; NCI; SPIKE; HumanCyc; Panther), the edges are extracted respectively. The edge sets of pathways are then created.
}


\format{An environment variable}
\author{Junwei Han <hanjunwei1981@163.com>, Xinrui Shi<xinrui103@163.com> and Chunquan Li <lcqbio@163.com>}
\keyword{file}